+++ src/aks-preview/HISTORY.rst
@@ -11,11 +11,12 @@ To release a new version, please select a new version number (usually plus 1 to 
 
 Pending
 +++++++
-* Vendor new SDK and bump API version to 2025-08-02-preview.
-* Pre-deprecate `--enable-custom-ca-trust` and `--disable-custom-ca-trust` in `az aks create`, `az aks update` commands.
+* Implement platform-managed-keys (PMK) awared validation for KMS customer-managed-key (CMK)
 
 18.0.0b44
 +++++++
+* Vendor new SDK and bump API version to 2025-08-02-preview.
+* Pre-deprecate `--enable-custom-ca-trust` and `--disable-custom-ca-trust` in `az aks create`, `az aks update` commands.
 * Hide `--enable-managed-system-pool` parameter for `az aks create` for now, as the feature is not ready yet.
 
 18.0.0b43